small improvements
Fourier descriptors work best when there is only one connected edge. To make this happen more often i applied a morphological close, an image fill, and another edge detection in order to only capture the outline of the object.

For the fft desciptors to not be a jumbled mess, the edge point indies need to be ordered such that if they were connected in a line, then they would reproduce the outline. To achieve this i tried sorting the indices based on the angle they are from the center of mass. This is not a perfect solution, especially if the shape is concave and the center of mass is outside the border.

The array matlab returns from the fft function does not contain the low frequencies in the beginning only. the low frequency data is at both ends of the array. To preserve the low frequencies, i ran fftshift to put the low frequenceise in the center, then i zeroed the edges, and ifftshift to put the data in the correct order again.
https://scicomp.stackexchange.com/questions/5314/order-of-matlab-fft-frequencies